src/domain/forum/application/use-cases/fetch-answer-comments.spec.ts
import { UniqueEntityID } from '@/core/entities/unique-entity-id';
import { makeAnswerComment } from '@test/factories/make-answer-comment';
import { InMemoryAnswerCommentsRepository } from '@test/repositories/in-memory-answer-comments-repository';

import { FetchAnswerCommentsUseCase } from './fetch-answer-comments';

let inMemoryAnswerCommentsRepository: InMemoryAnswerCommentsRepository;
let sut: FetchAnswerCommentsUseCase;

describe('FetchAnswerCommentsUseCase', () => {
  beforeEach(() => {
    inMemoryAnswerCommentsRepository = new InMemoryAnswerCommentsRepository();
    sut = new FetchAnswerCommentsUseCase(inMemoryAnswerCommentsRepository);
  });

  it('should be able to fetch answer comments', async () => {
    await inMemoryAnswerCommentsRepository.create(
      makeAnswerComment({ answerId: new UniqueEntityID('answer-1') }),
    );
    await inMemoryAnswerCommentsRepository.create(
      makeAnswerComment({ answerId: new UniqueEntityID('answer-1') }),
    );
    await inMemoryAnswerCommentsRepository.create(
      makeAnswerComment({ answerId: new UniqueEntityID('answer-1') }),
    );

    const result = await sut.execute({
      answerId: 'answer-1',
      page: 1,
    });

    expect(result.isRight()).toBeTruthy();
    expect(result.value?.answerComments).toHaveLength(3);
  });

  it('should be able fetch paginated answer comments', async () => {
    for (let i = 1; i <= 22; i++) {
      await inMemoryAnswerCommentsRepository.create(
        makeAnswerComment({ answerId: new UniqueEntityID('answer-1') }),
      );
    }

    const result = await sut.execute({
      answerId: 'answer-1',
      page: 2,
    });

    expect(result.isRight()).toBeTruthy();
    expect(result.value?.answerComments).toHaveLength(2);
  });
});
